add v3 client

+++ server/rest/user/users_service.py
@@ -1,5 +1,6 @@
 from db.models import BioGenomeUser
 from mongoengine.queryset.visitor import Q
+from errors import NotFound
 
 def get_users(offset=0,limit=20,
                 filter=None):
@@ -10,20 +11,27 @@ def get_users(offset=0,limit=20,
     return users.count(), users[int(offset):int(offset)+int(limit)]
 
 def create_user(data):
+    if not 'name' in data.keys():
+        return 'name is mandatory', 400
     username = data['name']
     ex_user = BioGenomeUser.objects(name=username).first()
     if ex_user:
-        return
+        return f'{username} already exists', 400
     new_user = BioGenomeUser(**data).save()
-    return new_user.to_json()
+    return f'{username} correctly created', 201
 
 def update_user(name,data):
     ex_user = BioGenomeUser.objects(name=name).first()
+    name = ex_user.name
+    if not ex_user:
+        return f'{name} does not exist', 404
     ex_user.update(**data)
-    return ex_user.to_json()
+    return f'{name} correctly updated', 201
 
 def delete_user(name):
     ex_user = BioGenomeUser.objects(name=name).first()
-    deleted_user = ex_user.to_json()
+    if not ex_user:
+        raise NotFound
+    name = ex_user.name
     ex_user.delete()
-    return delete_user
+    return f'{name} correctly deleted', 201
